+++ Labs/Lab 1/House.h
@@ -8,16 +8,16 @@ class House {
 		string owner;
 		string address;
 		int bedrooms;
-		float price;
+		long double price;
 
 	public:
 		House(){} // default contructor
 		void setOwner(string name);
 		void setAddress(string location);
 		void setBedrooms(int noOfBedroom);
-		void setPrice(float amount);
+		void setPrice(long double amount);
 		string getOwner() const;
 		string getAddress() const;
 		int getBedrooms() const;
-		float getPrice() const;
+		long double getPrice() const;
 };
